requirements.txt
streamlit
python-dotenv
pandas
scikit-learn
nltk
google-generativeai
